requirements.txt
fastapi
uvicorn[standard]
httpx
jinja2
cairosvg
pillow
python-multipart
google-genai
python-dotenv
requests
openai